Turn the logic of HEPMC_MINIMAL_WEIGHTS around as the default is to have minimal weight output only. Renamed to HEPMC_EXTENDED_WEIGHTS=<0|1> with default 0 now. As the Rivet interface has its own HepMC interface, the switch for the Rivet_Interface is renamed to USE_HEPMC_EXTENDED_WEIGHTS=<0|1> for consistency.
To be clear, to set it for the output to file set
HEPMC_EXTENDED_WEIGHTS, to set it for the Rivet_Interface set
USE_HEPMC_EXTENDED_WEIGHTS in its block. They can be used independently,
ie. full event info can be written to file and minimal event info given
to Rivet simultaneously.