Merge -c28696 ^/trunk/SHERPA .
Original commit message:

Add namespace specifications for instantiations

Add namespace to explicit class template instantiations, as otherwise clang
sees the omission as an error, if using

    CXXFLAGS="-std=c++11 -stdlib=libc++".

This enables compilation with clang linking to Rivet 2.5.0, which requires C++11.

For the same reason, an implicit cast has been made explicit and some weird
rvalue construction has been replaced with code that is readable and compiles.