Configure to use C++11 for Rivet v.2.5.0 and later
The reason is that C++11 is required by Rivet v.2.5.0 and later. This fixes the
compilation without the need for manually inserting '-std=c++11' or
'-std=gnu++11' into the CXXFLAGS.  Note that C++11 is now always used when
available on the build system. If so, the HAS_CXX11 prepocessor flag is set to
1, otherwise to 0.