Add comment on how to add command line options
Note that the new command line interface supports both the classic UNIX style
(e.g. -e) and the GNU style (e.g. --events), and is nearly POSIX-compliant. The
exception to the latter is that we allow to freely mix options and none-options
(e.g. "Sherpa -wWeighted EVENTS=1k" or "Sherpa ptmin:=50 -R 1")