Separate parton-shower emission variations
The ME-only reweighting results (excluding factors from the parton-shower
reweighting) can now be written to HepMC using the
HEPMC_INCLUDE_ME_ONLY_VARIATIONS setting.  If HEPMC_INCLUDE_ME_ONLY_VARIATIONS
is set to 1 (default: 0), these ME-only results are found in the HepMC output
with the key "ME_ONLY_<variation name>". The full variation output has not been
affected and is still found with the key "<variation name>".  This concludes
the work started in r31487 and r31539.