Enable the output of ME-only variations
This is enabled by setting HEPMC_INCLUDE_ME_ONLY_VARIATIONS to 1.  Also, the
treatment of scale variations in CSS and DIRE are aligned: The MC@NLO-like
one-step showers use the scale variation factors as the ME, whereas the
ordinary showers do that only when enabled (by setting
REWEIGHT_SPLITTING_ALPHAS_SCALES and/or REWEIGHT_SPLITTING_PDF_SCALES to 1,
corresponding to muR and muF, respectively).