Use blobs wgt instead of passed wgt in Outputs
That way, the structure is the same for Output_Base::Run as it is for
Analysis_Interface::Output, or on a higher level, for Output_Phase::Treat and
Analysis_Phase::Treat.

In the future, any parallel structure of passing weight in blobs and on its own
should be resolved, to prevent issues as the one fixed in r31547.