Check energies in CSS kinematics
This mitigates #44. A more proper check would be to check the kinematic
limits as described e.g. in arXiv:0709.1027. However, violations of
these limits will result in negative energies in most cases, such that
this commit uses an approximation which is good enough for the 2.2.6
release, whereas on master the limits will be enforced exactly after
merging the latest changes from the unlops branch.