Disable exit on Singlet_Checker pair mass check
This check had been introduced with 5ecdffec. It failed very quickly
e.g. for Examples/CI/LO_Z and thus made this test suite impractical to
find other issues.

I still print an error (albeit at a lower check accuracy, i.e. more
rarely), because the underlying issue might need to be fixed. I'll open
an issue for that.

This commit also lets the Taylor expansion in the kinematic limit for
gluon splitting kick in much less frequently. It's now more adapted to
the accuracy of the internal check of the gluon splitter,
Gluon_Splitter::CheckKinematics, such that we have much less fails of
this check.

You can see this in the following table. There is a sweet spot at using
the Taylor expansion when the kinematic ratios that are the small
parameters in the expansion exceed 1e6 (note that the kinematic checks
happen at an accuracy of 1e-6). Apparently, if not using the expansion
at all the numerical inaccuracy in the limit also leads to the kinematic
checker to declare failure. But using it too often obviously at some
point leads to the same, because then the error of the expansion gets
too large.

| Ratio at which the Taylor expansion is used | #kinematic fails | #taylor expansions |
|---------------------------------------------|------------------|--------------------|
| 8                                           | 4                | 0                  |
| 7                                           | 4                | 7                  |
| 6                                           | 1                | 74                 |
| 5                                           | 302              | 545                |

This was determined using the Examples/CI/LO_Z with 1k events and
`YFS_MODE=0`.

Note that using the Taylor expansion more rarely might lead to a
non-deterministic result for the event generation when comparing
different branches that should have the same result. This was the reason
to introduce the expansion in the first place, cf. #97. This needs to be
checked (and will be as part of validating !28).