Fix call to AlphaSLam with nonsensical scales
This fixes #102 on the caller side and is therefore preferred to the
previous attempt to fix this in 05749814, which is a workaround in the
wrongly called function.  This seems to be a safer solution to me.

I tested this patch by printing all members of p_thresh after the
initialization before and after the patch, and I get the exact same
results output (except for some uninitialized members that are never
set, i.e. the beta0 member for the AsDataSet for nf=0,1,2).