Handle YAML parsing exception properly
Also add a general std::exception catch clause in main() in case we miss
to catch a 3rd party exception somewhere else. Then we can at least
print it's description.

Note that the Sherpa constructor is now part of the main() try-clause,
because the YAML parsing is part of the construction and might raise an
exception.

Note also that rpa is now constructed before initializing the main
Settings instance, because rpa is involved in the exception handling
(it's responsible for writing citation info).