Improve the parsing of command line arguments
Before, each appearance of ':' was replaced by ': ' to allow for a more
compact notation, i.e. A:B instead of 'A: B' (note that the latter
example required an extra level of quotation). However, this did not
allow to give options that included ':' as part of their name or value,
in particular

  Sherpa 'FUNCTION_OUTPUT: {Sherpa::InitializeTheRun: 15}'

was not parsed properly due to the '::'. This is fixed now,
double-occurrences of ':' are ignored now.

In addition, when several ':' appear in an argument (but not in direct
succession), curly braces are inserted automatically to create a valid
YAML string, i.e. the argument of

  Sherpa FUNCTION_OUTPUT:Sherpa::InitializeTheRun:15

is now interpreted correctly as

  FUNCTION_OUTPUT: {Sherpa::InitializeTheRun: 15}

However, if you have several name-value pairs within a scope, you will
still need to use curly braces (and proper quotation), e.g.

  Sherpa 'FUNCTION_OUTPUT: {Sherpa::InitializeTheRun: 15, Sherpa::~Sherpa: 8}'