Reorganise selector parsing to fix a crash when default differ
The crash appeared when two selectors set defaults for the same
subsetting, e.g. like this:
```
SELECTORS:
- Type: a
  Some_Param: x
- Type: b
  Some_Param: y
```
... and we have different defaults for Some_Param for a vs. b. The
exception is thrown because a previously set default would otherwise be
modified (since they were not scoped properly). Now selector settings
are supposed to be given in the following way:
```
SELECTORS:
- a:
   Some_Param: x
- b:
   Some_Param: y
```
This scopes the different Some_Param settings (and their defaults) in
the right way (i.e. under a and b), and follows the DRY principle,
because the Type boilerplate is now removed.

All examples are adapted.