Remove shadowing global declarations in SWIG interface
For Python 2.7.1, SWIG 4.0.0 and clang-1001.0.46.4 on macOS the previous
way to (re-)declare global instances led to various parallel
declarations, s.t. the Random instance set up in Sherpa.C was not
visible later when running Run_Parameters::Init, where it was still
NULL, which led to a crash. This has been fixed for the above setup, and
has also been tested for python 2.7.15, SWIG 3.0.12 and g++ 8.3.1 on
GNU/Linux, where it does not change the behaviour.