Update Rivet_Interface for Rivet 3.0.0
This adds a separate implementation of Rivet_Interface if Sherpa is linked against Rivet3. The old interface for Rivet2 with many additions for the weight handling is still available if Sherpa is linked against Rivet2.

Eventually we can remove Rivet2 support by deleting the corresponding part of Rivet_Interface.C.