Add synonyms to refer to defaults of Settings
This could e.g. be used to introduce a new magic value such as "default"
that can be used to set a Setting explicitly to its default value
(potentially overriding previous non-default choices).  The current
implementation is not used yet. It allows for a more fine-grained choice
(i.e. default synonyms can be set for each Setting individually).