Update DiHiggsNLO examples to new NLO syntax
The NLO process order definition has been updated to the new standard.
In order to be able to validate the examples, in both files
BEAM_REMNANTS has been set to false since it currently causes an
infinite loop due to an unrelated issue.

Note that the OpenLoops-based example does not work yet. This will be
resolved in #205.