Propagate recent parsing changes for SELECTORS also to PROCESSES
The SELECTORS parsing had been changed in
2e61d46c19e9e35545a2760641150a949ccfe861 to fix a crash. For the
PROCESSES parsing, a similar crash does not appear (because all
processes share the same defaults). However, it's better to use the same
specification pattern for the settings, to reduce surprises. Also it's
easier to read and removes boilerplate.

E.g.
```
PROCESSES:
- Process: a b -> x y
  Some_Param: x
- Process: c d -> x y
  Some_Param: y
```
are now written as follows:
```
PROCESSES:
- a b -> x y:
   Some_Param: x
- c d -> x y:
   Some_Param: y
```

This change has been tested with the CI/LO_Z example, also by adding a
second process as in the example snippet above.