Use Status=11 for incoming particles with a strong charge for HepMCShort
Before those had Status=4 (as all incoming particles). But this led to a
crash in Rivet (as reported in #174), because Rivet does not recognise
partons as proper beam particles (which makes sense physically). The
workaround was to use Rivet with its --ignore-beams option.

Status=11 is a reserved status for MCEG and is the default fallback in
`Sherpa*HepMC(ATOOLS::Particle*, HepMC::GenParticle*&)`, which was the
reason for this choice.

(cherry picked manually from commit 266eaefb004acf80d2931aed88a8d2476565dc34)