Generalise Sudakov ampl mom updating for boosted initial states
The previous algorithm made the assumption that the initial state
momenta are in the rest system of the initial state. This was fine for
the leptonic IS tested so far (e.g. for the checks against
Denner/Pozzorini), where both beams had the same energy, but led to
gigantic K factors as soon as we looked into pp initial states
(e.g. pp -> WW).