Add a hacky solution to write out individual EW Sudakov contribs
... via the HepMC2 and Rivet interfaces. Setting KFACTOR=EWSudakov no
longer applies anything to the nominal weight, instead the data for each
contrib is filled into the ME wgt info object and then later in the
HepMC and Rivet interfaces the information is either applied to the
nominal weight or written out as individual YODA files (where the sum
should give K*nominal again). This behaviour can be controlled for YODA
output using the new Rivet SPLITEWSUDAKOVCONTRIBS option (default:
true).