Experimentally disable COMIX process mapping
... since otherwise Kfactors are only set up for the unmapped processes,
such that we do would not operate on the right flavours. Also, this
prevented the Kfactor contributions stored in the ME_Weight_Info object
from being properly propagated to the Rivet interface (since the
information was stored in the unmapped process' information, not in the
mapped one's).

We might want to replace this later with a proper fix. Also, we
definitely not want to merge this into master!