Make Fusing_Direct_Hook work with multiple processes
Currently the `Fusing_Direct_Hook` is restricted to setups with one process (group), which is not sufficient if we want to generate e.g. W+ and W- in one setup. This is generalised in this commit.

(cherry picked from commit e09e1faa1be2a44fe976d0e60ad4d0e5555fbd2a)