Re-enable flavour ordering (and EWSudakov amplitude caching)
... since it led to wrong cross sections even when not using the
EWSudakov KFactor at all. The new paradigm is that the EWSudakov module
is not flavour-sorting anything as far as possible to keep book-keeping
simple. Only when accessing the ME calculation via the Comix_Interface
is the ordering is applied.

The amplitude caching is also re-enabled and now taken care of by the
Comix_Interface objects. It must be carefully checked that this caching
is correct, i.e. reset whenever needed, such that calculations take
place when expected.