Spin correlations for taus from perturbative decays
This implements a workaround to recover spin correlations for taus from hard decays, e.g. top(->W(->tau(->pi nu) nu)b). It's limited in the sense that it will only take into account the tau production polarisation in a given tau decay, but not correlations between the tau decay and other decays across the decay cascade. But this is already giving the most relevant effects.

A proper and fully general implementation of this is pursued in parallel on the unified-decay-handling branch which is targetting Sherpa 3.0.

(cherry picked from commit 9144d5e899d7c649b6bbbeb76a2b93e9259fda86)