Impose Breit-Wigner only for unstable particles
This used to be option MASS_SMEARING=2, while the default (=1) would also shift stable particles offshell if they have a width. The latter is not implemented anymore.

Several bugfixes for mass smearing, e.g. where DecayBlob() was used instead of !Stable(), which is not useful anymore in the new shower handling where particles might have shower blobs as DecayBlob().