read form factors from UFO model
Final step for porting MadGraph form factors syntax (https://cp3.irmp.ucl.ac.be/projects/madgraph/wiki/FormFactors) to Sherpa.

This generally allows using model parameters inside of the Lorentz structure definition. Lorentz structure is a natural place for the formfactors because the latter might depend on scalar products of the four momenta, which are handled mostly inside of the lorentz structure module.