allow elementary operations with elementary tensors (tensor vs number)
Here is an improvement to UFO model translator which allows for division by elementary tensors, addition and subtraction between elementary tensors and numbers.

Example of what is wanted: ~1/(1 - a^2/(P(-1, 3)*P(-1, 3)))^2, having scalar factor inside of the vertex which allows dependency on scalar products of momenta.

Note: In master we already have this functionality. It is implemented with sympy, while in rel-2-2-9 there is no sympy yet and we use c_variable for this. This commit is to implement division by elementary tensor with c_variable.