Add a hack to account for M^phi- = - M^W-
a factor which had been missing in the equations of our primary EWSud
reference arXiv:hep-ph/0010201, but which can be found e.g. in the
associated PhD thesis arXiv:hep-ph/0201077 and according to Davide also
in textbooks.

The hack only fixes the test case eeWW (LL) in the coefficient checker
of the EWSud module. It needs to be properly implemented to be correct
in general.