Use the physical phase to evaluate ME ratios (except for PR coeffs)
The idea was to see if this solves the eeWW (LL) coefficient issue when
comparing to the ref. Unfortunately, the reference is only reproduced
when swapping the sign of Ipm(phi) -> chi, but I don't see a reason why
that should be the case. So there may be some other issue still.