Fix couplings used for EWSud calculations to conform to Sherpa's
Previously, they followed (mostly) the conventions by Denner and
Pozzorini [arXiv:hep-ph/0010201] (except for the sign of \sin\theta_W,
which we flipped to account for a difference in the definition of the
covariant derivative, but that was not enough to get everything right).
In particular the couplings in the neutral gauge boson sector (including
the neutral Goldstone boson \chi) followed different conventions. This
led to wrong coefficient calculations when combined with ME ratios
calculated by Sherpa.

Note that all ME ratios are now calculated in the Goldstone boson
phase (except for those where a we can be sure that longitudinal
boson legs are not affected).