re-enabled PR calculation
due to the UFO being auto generated, couplings were not correctly
initialised when using the high energy model. Created a local map of
complex constants that contains only quantities that are derived from
the base p_complexconstant (which is the one we update with the
highenergy couplings), which is cleared at every call.