Clean up GBET conversion and leave a comment to note that it fails
The reasons are unclear at the moment. However, calculating the ME
ratios in the Goldstone phase for the EWSud coeffs seems to give the
right results, confirmed by all coeff checks that we do. Hence fixing
this does not seem important at the moment. Also see the comment left
in EWSudakov_Calculator::GBETConversionFactor.