Fix crash in setups with neutron beams
In setups with neutron beams the neutron was initialised in the beam handler, not in Ahadic. This initialisation was using a wrong spin 0 (https://gitlab.com/sherpa-team/sherpa/-/blob/rel-2-2-9/BEAM/Main/Beam_Spectra_Handler.C#L390). This commit changes it to be the same as in Ahadic (https://gitlab.com/sherpa-team/sherpa/-/blob/rel-2-2-9/AHADIC++/Tools/Hadron_Init.C#L272). When encountering decays that involve neutrons and have matrix elements implemented, this would otherwise lead to crashes because of a mismatch of the expected number of helicity combinations in the ME.