Bugfix for crash in ReweightWithoutSubevents
In setups with OTF shower variations when all shower emission scales are below the shower reweighting cutoff the run will crash. This is because the reweighting tries to access the last element of the alphas reweighting factors, which does not exists in those cases. A safeguard preventing such an access is now implemented.

This bug does not lead to wrong results, only a crash. It exists in the code in this form since v2.2.6.