Fix MCATNLO shower reweighting
In the case of no splitting, the m_last pointers must be set to NULL (as
in CSS; note that the way DIRE/DIM is operating the OTF reweighting
differs, such that the we do not need to port the fix to those), such
that the Reweight function uses the right placeholder for the
non-existing splitting scale (i.e. 0). Otherwise it potentially stops
reweighting early, re-using some outdated splitting scale from an
earlier shower evolution (actually, the bug was even non-deterministic,
such that the m_last were probably not pointing to valid memory
anymore).

I have tested that the nominal outcome of a few hundred events is
exactly the same as before in MC@NLO Z production, and that the
non-deterministic outcome of the OTF reweighting is replaced with a
deterministic one (which seems to be correct).

Note that the cases where I have seen the issue, the effect on the
overall event weights was at a sub-percent level, and it affected only
O(percent) of all events.