Make new Sherpa 2.2 ref more cross-compatible
Having a colon (":") in the reference key caused an issue for me
compiling the Manual; this might depend on the texi version etc.
I also fixed the data field for this bibitem, such that it is properly
typeset in the reference list.

I have also updated the single-top reference with a journal entry, it
was preprint-only before.

(cherry picked from commit 9c951143d239363772b9de0ee012fcc5d6be49f0)