Fix Rivet path handling for MPI runs
Fixes #261.

MakeDir should be called by all ranks, not only the zero rank, since
MakeDir is responsible for the "MPI reduction" (e.g. not calling mkdir
for each rank, but only once.)
If MakeDir is only called by the zero rank, Bcast within MakeDir is not
called by all ranks, which will lead to a crash at a later point.