Add missing check when setting K factors in AMEGIC
Fixes #264.

K factors (at least the variable K factor) assume that the scale setter
of the process is non-NULL. There was a violation to this rule in
`AMEGIC::Single_Real_Correction::SetKFactor`, because of a missing check
for `m_no_tree` (which was only present in `SetScale`, but not in
`SetKFactor`, thus leading to a situation where the latter is called for
the tree process, but not the former). When using a variable K factor,
this led to an exception during process set-up.


(cherry picked from commit 606e8b37520d913c6d8926b242506b851acc4fde)