removed were possible user-defined lorentz structure, for amegic
This commit re-enables the use of a SMGold coupling, such that
goldstone boson verts are disentangled from standard EW ones (and no
doublecountings are possible). In addition, removes the references to
the lorentz functions automatically defined by the UFO, such that
AMEGIC can also be used. Note, however, that you would still get
warnings that some verts have not been found, like FFSL/R, VVVV and
VSS. This means that with AMEGIC + SMGold is not possible to generate
procs with goldston bosons or that have quartic-gauge couplings (for
the moment). Fixe and close #4