Allow for tags in process specifications
Tags are not automatically resolved in setting keys, hence this is now
explicitly done in this case, to allow (again) for process
specifications such as "93 93 -> 11 -11 93{$(NJET)}".

This has become necessary because 8cb7d418af made the process
specifications into the key with its value being the process setting
map (instead of the spec just being another value of an entry in the
process setting map).