Require the SMGold cpl to be set explicitly
... when using it. This is to make sure that AMEGIC::Amplitude_Handler
does not receive inconsistent lengths of coupling vectors from the
process initialisation and the vertices provided by SMGold. This happens
because per default the Process_Info only initialises coupling vectors
of length 2, such that an omitted Order specification in the Process
settings will keep that default, while SMGold operates with 3 orders. It
is unclear how to solve this best at this point, but requiring an
explicit setting should be safe for now. It is also unclear why COMIX
did not complain about this.