reverted to rel behaviour for the evaluation of PDF in the shower.
In particular this commit uses the CMW scale for the evaluation of
PDF as well as alphaS in the shower, implicitly raising the shower
cutoff. This fixed the soft-enhancement behaviour seen in comparisons
to release.

NOTE: this would need to be complemented with tuning and possibly the
re-introduction of KPERP in the remnants.