Fix for interplay of Enhance_Function/Observable with Max_Epsilon treatment
The Max_Epsilon weight histogram is used for a dynamic setting of the unweighting maximum. Since the unweighting has to reject against a maximum that includes the enhancement, the weight histo should not be dividing the enhancement factor out.

This was cured/masked by the faithful overweight treatment, but would lead to larger weight fluctuations when the Enhance_Function/Observable is used.