Apply missing flavour swap when filling Weight_Info
Before, the flavour information became out of sync with the Bjorken x,
pdf scale and pdf value information, when a swap was applied.

The reason was probably that the flavour information is taken from the
active process, whereas all the other information originated from the
ISR handler, which might have already applied the swap to the
information it offers.