Sanitise HepMC weight parsing in the Rivet interface
... within the workaround of a linked HepMC library without working
named weights, i.e. HEPMC_HAS_NAMED_WEIGHTS=1 and
HEPMC_HAS_WORKING_NAMED_WEIGHTS=0.

This is not a fix, because the parsing as implemented prior to this
commit worked, but only accidentally (because the >> operator of string
streams ignored the closing ')', which was included in the substring
that was supposed to contain only the number before the ')'). However,
now the intent is clear.