Fix for Internal1 error in pp->ttll
Do not contract taus in the original Amplitude2_Tensor as presumed "stable" particles during the hard decay handler phase. This way their polarisations are not available anymore later for the hadron (+tau) decay handler phase, leading to an Internal1 error. Now tau polarisations (if unstable) are kept in the Amplitude2_Tensor.