Print a warning when settings have not been read
This will be printed at the end of the run. More details about unused
settings can then be found in the `Settings_Report` created by the
`Settings_Writer` class.

A boolean setting `CHECK_SETTINGS` has been added (default: `true`). If
it is set to `false`, the Settings_Report will not be created, and the
check for unused settings (which is implemented in `Settings_Writer`) is
disabled.