Replace `dependencies` -> `needs` in .gitlab-ci.yml file
It should be functionally the same (but see below), but the Visualisation of the dependency graph within GitLab only works with the more recent keyword `needs`, it seems.

One functional difference might be that with `needs` jobs can start before the previous stage has fully completed, such that for example the clang test run can start immediately after the clang test build (instead of waiting also for the gcc test build). This might save some time.