Fix norm factor in assew weight calculation
The previous commit re-ordered the assew weight calculation and the
application of m_dadswgtmap to m_last in Single_Process. But the
normalisation factor used for the assew weight calculation assumed that
the m_dadswgtmap had already been applied. Now, the nominal DADS weight
is explicitly subtracted from the normalisation factor.