Make custom Model example Makefiles more robust
... for compilation with clang. We need to use the linker flags
"-undefined" and "dynamic_lookup", otherwise just linking with
libSherpaMain (where the required symbols are usually not defined) will
give a linker error. Using those flags, not finding a symbol will just
be ignored, with the promise that they will be available during runtime
(which should be true in these cases). This seems to be the default
behaviour for other compilers (gcc) in this case.