Specialise ToType<double> to treat errors more manually
To be precise, in the generic implementation using string streams,
underflow leads to fail() returning true with a recent version of clang,
while for an older version of gcc it returns false.
Hence the specialisation uses strtod instead, and checks for range
errors explicitly. Those are now only reported, but do not result in an
exception to be thrown, as in the generic implementation.