Suppress confusing HepaHepMC[3]Output lib loading warning
When ANALYSIS=Rivet is used, the init handler tries to load at least one
HepMCOutput library. Before, each failed attempt would print a warning
about the failure.
Now, an attempt is only made when the HepMCOutput library is actually
built, such that there is no futile attempt in the first place, such
that a confusing warning about a library that the user did not even want
to be built is not issued.