Apply EWSUDAKOV_MAX_KFACTOR directly to EWSudakov->KFactorExp
Before, the maximum was only checked for KFactor itself, and KFactorExp
was calculated from this capped KFactor (such that it ends up being
potentially much larger than the maximum, up to exp(maximum) in fact).
Now, the condition is checked individually for KFactor and KFactorExp.

As for the individual contributions, they are not longer affected by the
maximum, i.e. they are always left as they come out of the calculation.
That means that KFactor = 1 + \sum_i contrib_i is no longer true.
Before, this was always true, because the contributions were set to 0.0
when KFactor was set to 1.