Allow for cmd line input such as `A:B:X A:C:Y`
Before, the duplicate top-level key `A` would have prevented the first
argument `A:B:X` to have any effect, because YAML does not allow for
duplicate keys in a mapping. This is highly unexpected for the user.

Now, both arguments have an effect. This is achieved by supporting YAML
multi-document files, and handling each command line argument as a YAML
document. Thus, the unique-key rule does not apply, and we can traverse
the YAML documents when we try to read a given setting until we find it.