Fix crash when handling PROCESSES from multiple config files
PROCESSES is a list of mappings, where each mapping can only contain one
key, the process definition string. The process definition string was
read using Settings::GetKeys(). But this method used to merge keys from
all config sources. To understand the issue, consider the following
example:

file 1:
PROCESSES:
- "x y -> a b": {}

file 2:
PROCESSES:
- "x y -> a b c": {}

In this situation, GetKeys() returned *both* process definition string,
thus triggering the one-key guard, leading to a crash.

To prevent this, we now disable key merging as soon as we are within a
sequence. This makes sense, because we would not be able to map elements
from two sequences in a meaningful way, as the above example proves.