Setup EWSud to work as a generic Hard_Process_Variation_Generator
Note that the initialization of EWSud is now triggered by
InitializeTheReweighting in the Init Handler. Before it was done
implicitly via the initialization of the Matrix Element Handler.
The advantage is that the reweighting is only set up if needed (i.e. if
more than one event is to be generated).
This can save a lot of unnecessary init time in a pure integration run.

Note also that the old structures are still in place. They still need to
be deleted, and the EWSud test cases need to be adapted.