Add --enable-ewsud, add EWSud weight names facility
The latter will be used by the HepMC and/or Rivet interfaces for
translating the Weights Map to HepMC weight names.

Note that I have loosened the EWSud coefficient checks a little when the
reference values happens to be close to zero. It is sufficient then to
be within 0.05 of that value.