Add HIGGS_INTERFERENCE_OQCD_OVERESTIMATE setting (default: 0)
Using HIGGS_INTERFERENCE_OQCD_OVERESTIMATE=1 makes the Higgs_Tree
process to return 3 when calculating mixed background-signal terms where
the terms are of different QCD orders 2 and 3. The on-the-fly
reweighting will use the value of 3 to calculate renormalisation scale
variations, thus overestimating the scale uncertainty.

The default value of 0 retains the previous behaviour, i.e. Higgs_Tree
returns 99 in the mixed case, as a nonsense value. The reweighting will
then throw an exception.