Revert making the Init Handler a global object
This was only a hack that allowed to create another HE version of the
model later in the EWSud module. This is now enabled otherwise, by
copying a pointer to the ISR handler map within Model_Base, such that
all information is present within a model to generate a copy of it.
The init handler is now again a member of the Sherpa class, as it was
previosuly.