Improve reporting on HepMC settings inconsistent with variation weights
Namely, when `HEPMC_USE_NAMED_WEIGHTS: false` (the default), but the
user tries to use variation weights.
In addition, this is now also flagged when a userhook is used, since
this is also only written out when named weights are enabled.

Also, we do not throw any longer, but just print a warning once, given
that the user might only be interested in the total cross section of a
variation, which can be read off the Sherpa output.