Add EWSud AddOn and auxiliary SMGold Model
Resolves #11.

For more details on the details of the implemented EW Sudakov
(= EWSud) calculations, see arXiv:2006.14635 and 2111.13453.

SMGold is the Standard Model extended by Goldstone bosons,
which are used by the EWSud add-on to calculate corrections.

Use --enable-ewsud when configuring Sherpa to enable the add-on.

EWSud corrections can be enabled for the nominal calculation via
`KFACTOR: EWSud` or as alternative event weights using

    VARIATIONS:
    - EWSud

For additional details, see the manual section "Approximate Electroweak
Corrections", which now discusses EWSud and EWVirt corrections
collectively.

Note that this introduces the `make check` target. For now, only the
EWSud AddOn has checks, which validates the EWSud calculation using
various known literature results. The CI is configured to run this check
as part of the pipeline.

This is a squashed commit for !464, the detailed commit history is
recorded in the 11-ew-sudakov-logs+commit-history branch, and consists
mostly of commits by me and Davide Napoletano. Please contact one of us
in case of regressions introduced by this commit.