Fix overload resolution for InitAmplitude in AMEGIC
I have deleted an unused int argument of InitAmlitude in
Single_LOProcess_External and Single_LOProcess_MHV. However, I have not
changed the caller side, and because the function has an overload with
std::vector<double>* instead of the int, calls were actually resolved to
the second one (because int implicitly converts to the pointer!). This
potentially led to an inconsistent setup of the dipole subtraction in AMEGIC.

An example is the setup of BVI in dijet production, where the wrong
InitAmplitude was used when initialising MHV amplitudes.