Fix bug in pythia8 interface from multiple partonic decays.
When sherpa handles the decays of unstable hadrons from pythia8
hadronization some particles decay partonically. A previous fix
for a problem when pythia8 handles the decays itself caused only
the first of these partonic decay to be hadronized again.
This change makes it that the previous change is now only applied
when pythia8 handles decays and not when sherpa does it.