Allow for rerun/pilot run deviations, but print ERROR
The reason is that I am not sure how large deviations from numerics
might be, and in that case we do not want to crash, but rather print an
error message and continue. At the moment, deviations of 1e-6 in the
Weight_Info objects will result in an error message.

(cherry picked from commit d8fced5a7fa77432a248e0f4ba962ac84f5b1052)