Fix crash when summing weight maps of events of different type
Sometimes an event has different entries in its weight map compared to
other events (e.g. the QCUT key is not always present). In that case,
some additional care must be taken when adding weight maps (e.g. to
build the sum of weights in the event handler). This is now handled
correctly in the Weights Map class and the crash is fixed.