Fix assignment of swapped initial states to remnants
Due to symmetry reasons, the initial states are swapped if the PDFs allow for it, see ISR_Handler.C for the code and ask Stefan for the details.
If such a swap happened, the correspondence of indices between particles and remnants is broken. That means that we have to check whether a swap has happened by looking if the particles follows the usual coordinate convention.
If the convention is respected, no swap has happened and particle[0] corresponds to remnant[0].
If the conventions is violated, the IS were swapped and particle[0] corresponds to remnant[1].
Thus we ensure that no events are rejected only because we extract a particle from the wrong remnant.