Avoid mapping error for massless initial states
For yexp = 0.999, the Simple_Pole_For/Backward channel is not invertible, due to an appearing exponent of 10**-3
In case of massive beams, e.g. protons, this justo a  rejection, as the points will violate the mass constraint.
However, for massless beams, e.g. photons, the points don't get rejected, and a mapping error appears as the random number cannot be reconstructed in the WeightYBackward function.
With this hack, we simply don't use the respective channel that leads to this error.