Reintroduce [PDF|SCALE|QCUT]_VARIATIONS settings
The VARIATIONS syntax is great to specify correlated variations (and can
still be used without change). But varying scales and PDFs etc. at the
same time is arguably a very rare case. The standard use case is
independent variations, and the verbose VARIATIONS syntax is a bit
awkward then. Hence, one can (again) use the more compact settings as
specified above. Examples would be

  SCALE_VARIATIONS:
  - [0.25, 0.25]
  - [4.00, 4.00]

... or alternatively, for a 7-point scale variation:

  SCALE_VARIATIONS: 4.0*

Similarly for PDF and QCUT variations:

  PDF_VARIATIONS:
  - NNPDF30_nlo_as_0118*
  - CT10nlo
  - CT10nlo/1

  QCUT_VARIATIONS:
  - 0.5
  - 2.0